% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildScenarios.R
\name{BuildScenarios}
\alias{BuildScenarios}
\title{Function to build scenarios.}
\usage{
BuildScenarios(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{BuildScenarios} builds structure to run mulitple scenarios.
}
\details{
This function builds scenarios from scenario input levels. The folder names
in the scenario input folder is used for naming scenarios. The function
creates folder for each scneario, copy the model files to the scenario folder,
and replaces the model files with relevant scenario inputs.
}
