% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunScenarios.R
\name{getNumberOfRunningTasks}
\alias{getNumberOfRunningTasks}
\title{Function that returns the number of current future tasks.}
\usage{
getNumberOfRunningTasks(asyncTasksRunning = vector(mode = "list"))
}
\arguments{
\item{asyncTasksRunning}{A list of asynchronous tasks running currently.}
}
\value{
An integer indicating the number of tasks currently running
}
\description{
\code{getNumberOfRunningTasks} returns the number of tasks running
currently.
}
\details{
This function returns the number of tasks running currently.
}
