% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunScenarios.R
\name{getRunningTasksStatus}
\alias{getRunningTasksStatus}
\title{Function that returns the status of tasks running currently}
\usage{
getRunningTasksStatus(asyncTasksRunning = vector(mode = "list"))
}
\arguments{
\item{asyncTasksRunning}{A list of asynchronous tasks running currently.}
}
\value{
A string indicating the status of the tasks running currently
}
\description{
\code{getRunningTasksStatus} returns the status of tasks running
currently.
}
\details{
This function returns the status of tasks running currently.
}
