% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSPMResults.R
\name{getScenarioResults}
\alias{getScenarioResults}
\title{Function to read all the results from a scenario directory
\code{readTables} reads all the model run results from a specified
directory.}
\usage{
getScenarioResults(ScenarioPath, ...)
}
\arguments{
\item{ScenarioPath}{A string identifying the path in which model files are
stored.}
}
\value{
A data.table containing all the tables.
}
\description{
This function takes the model path and returns a data.table containing
all the results. The additional arguments are passed to \code{readTables}
function.
}
