% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewScenarioSet.R
\name{NewScenarioSet}
\alias{NewScenarioSet}
\title{Function to build scenarios.}
\usage{
NewScenarioSet(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{NewScenarioSet} builds structure to run mulitple scenarios.
}
\details{
This function reads the scenario management database and creates
the related configuration json files for defining scenario sets.
It also creates all the scenario folders with their required inputs 
in the root directory
}
