% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewScenarioSet.R
\docType{data}
\name{NewScenarioSetSpecifications}
\alias{NewScenarioSetSpecifications}
\title{Specifications list for NewScenarioSet module}
\format{
A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{scenario input data to be loaded into the datastore for this
 module}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}
}
\source{
NewScenarioSet.R script.
}
\usage{
NewScenarioSetSpecifications
}
\description{
A list containing specifications for the NewScenarioSet module.
}
\keyword{datasets}
