% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunScenarios.R
\name{processRunningTasks}
\alias{processRunningTasks}
\title{Function that processes the tasks running currently}
\usage{
processRunningTasks(asyncTasksRunning = vector(mode = "list"),
  wait = FALSE, catchErrors = TRUE, debug = FALSE,
  maximumTasksToResolve = NULL)
}
\arguments{
\item{asyncTasksRunning}{A list of asynchronous tasks running currently.}

\item{wait}{A logical. TRUE if the processor needs to wait for all the results.}

\item{catchErrors}{A logical. Set to TRUE if want to catch the errors resulted
in futures.}

\item{debug}{A logical. Set to TRUE if need to print intermediate results.}

\item{maximumTaskToResolve}{An integer for the maximum number of tasks to resolve.}
}
\value{
An integer indicating the number of tasks currently running
}
\description{
\code{processRunningTasks} processes the tasks running
currently.
}
\details{
This function is called periodically, this will check all running asyncTasks
for completion. Returns number of remaining tasks so could be used as a boolean
}
