% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSPMResults.R
\name{readTables}
\alias{readTables}
\title{Function to read tables from given scenario
\code{readTables} reads tables from a specified scenario.}
\usage{
readTables(Year = NULL, Scenario = "Base", Output = "All",
  Table = TRUE)
}
\arguments{
\item{Year}{A string indicating the year for which the output needs to
be obtained.}

\item{Scenario}{A string for name of the scenario to be read into.}

\item{Output}{An array of character. If specified as 'All' then
all the tables are read in.}

\item{Table}{A logical indicating whether the 'Output' option contains
variable names or table names. Defaults to TRUE indicating table names.}
}
\value{
A data.table containing all the specified tables.
}
\description{
This function takes scenario name, a list of tables and read them into
data.table format and returns a list of data.tables.
}
