% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunScenarios.R
\name{RunScenarios}
\alias{RunScenarios}
\title{Function to run scenarios.}
\usage{
RunScenarios(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{RunScenarios} runs all scenarios.
}
\details{
This function runs scenarios in parallel. It uses the future package and
run scenarios asynchronously.
}
