% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioResults.R
\name{ScenarioResults}
\alias{ScenarioResults}
\title{Function to gather scenario results.}
\usage{
ScenarioResults(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{ScenarioResults} gather scenario results.
}
\details{
This function gather results from scenarios asynchronously.
}
