% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunScenarios.R
\name{startAsyncTask}
\alias{startAsyncTask}
\title{Function to start asynchronous task.}
\usage{
startAsyncTask(asyncTasksRunning = vector(mode = "list"), asyncTaskName,
  futureObj, callback = NULL, debug = FALSE)
}
\arguments{
\item{asyncTasksRunning}{A list of asynchronous tasks running currently.}

\item{asyncTaskName}{A string indicating the name of the task.}

\item{futureObj}{An object of class future.}

\item{callback}{A post processing function when the future returns
a value.}

\item{debug}{A logical. TRUE if want to print more intermediate messages.}
}
\value{
A list containing all the parameters to the function.
}
\description{
\code{startAsyncTask} calculates performance metrics.
}
\details{
This function starts/queues a future task.
}
