% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERPATResults.R
\docType{data}
\name{VERPATOutputConfig}
\alias{VERPATOutputConfig}
\title{Scenario output configurations for VERPAT model}
\format{A list containing configuration for models in json format:
\describe{
 \item{VERPAT}{Scenario output configuration for VERPAT model}
}}
\source{
VERPATResults.R script.
}
\usage{
VERPATOutputConfig
}
\description{
A list containing output configurations for VERPAT model.
}
\keyword{datasets}
