% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSPMResults.R
\docType{data}
\name{VERSPMOutputConfig}
\alias{VERSPMOutputConfig}
\title{Scenario output configurations for VERSPM model}
\format{A list containing configuration for models in json format:
\describe{
 \item{VERSPM}{Scenario output configuration for VERSPM model}
}}
\source{
VERSPMResults.R script.
}
\usage{
VERSPMOutputConfig
}
\description{
A list containing output configurations for VERSPM model.
}
\keyword{datasets}
