% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewResults.R
\name{ViewResults}
\alias{ViewResults}
\title{Function to view scenario results.}
\usage{
ViewResults(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{ViewResults} shows the scenario results in a html page.
}
\details{
This function shows the results of scneario run in a html page.
}
